Add Typescript's types to Pipeline Source page
Move Pipeline Source page to typescript
